+++ GUI/Search/OptionsMenuButton/OptionsMenuButtonMixin.lua
@@ -35,10 +35,10 @@ function KrowiAF_SearchOptionsMenuButtonMixin:OnMouseDown()
 					IsTitle = true
 				});
 
-	self:AddCheckBox(menu, addon.L["Search IDs"], addon.SearchOptions.db, {"SearchIds"});
+	self:AddCheckBox(menu, addon.L["Search IDs"] .. " (#)", addon.SearchOptions.db, {"SearchIds"});
 	self:AddCheckBox(menu, addon.L["Search Names"], addon.SearchOptions.db, {"SearchNames"});
 	self:AddCheckBox(menu, addon.L["Search Descriptions"], addon.SearchOptions.db, {"SearchDescriptions"});
-	-- self:AddCheckBox(menu, addon.L["Search Criteria"], addon.SearchOptions.db, {"SearchCriteria"});
+	self:AddCheckBox(menu, addon.L["Search Criteria"] .. " (@)", addon.SearchOptions.db, {"SearchCriteria"});
 	self:AddCheckBox(menu, addon.L["Search Rewards"], addon.SearchOptions.db, {"SearchRewards"});
 
 	menu:Toggle(self);