+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 - Toggle window once opened option
 - Option for localization team to localize time formats for event reminders
 - Option (on by default) to open the search results in a category instead of the inline window
+- Searching criteria can now be done without prefix if checked in the options
+- Ability to remove cached characters through the data manager window
 
 ### Fixed
 - Clicking the World Map Button will now properly update the category when it's already selected